Splitting tensor_ops/binary_map into separate files
