+++ src/tensor_ops/mod.rs
@@ -93,43 +93,53 @@
 //! ```
 
 mod arith_scalar;
-pub mod binary_map;
 mod broadcast;
+mod impl_add;
 mod impl_backward;
 mod impl_clamp;
+mod impl_div;
 mod impl_dropout;
 mod impl_mask;
 mod impl_max_axis;
+mod impl_maximum;
 mod impl_mean;
 mod impl_mean_axis;
 mod impl_min_axis;
+mod impl_minimum;
+mod impl_mul;
 mod impl_nans;
 mod impl_normalize_axis;
 mod impl_softmax;
 mod impl_std_axis;
+mod impl_sub;
 mod impl_sum;
 mod impl_sum_axis;
 mod map;
 mod matmul;
 mod reduce;
 mod select;
-mod utils;
+pub(crate) mod utils;
 
 pub use arith_scalar::*;
-pub use binary_map::*;
 pub use broadcast::*;
+pub use impl_add::*;
 pub use impl_backward::*;
 pub use impl_clamp::*;
+pub use impl_div::*;
 pub use impl_dropout::*;
 pub use impl_mask::*;
 pub use impl_max_axis::*;
+pub use impl_maximum::*;
 pub use impl_mean::*;
 pub use impl_mean_axis::*;
 pub use impl_min_axis::*;
+pub use impl_minimum::*;
+pub use impl_mul::*;
 pub use impl_nans::*;
 pub use impl_normalize_axis::*;
 pub use impl_softmax::*;
 pub use impl_std_axis::*;
+pub use impl_sub::*;
 pub use impl_sum::*;
 pub use impl_sum_axis::*;
 pub use map::*;
